+++ python/tests/test_discover.py
@@ -5,172 +5,89 @@
  * All rights reserved.
 """
 
-import requests
-import ouster.cli.plugins.discover
-from ouster.cli.plugins.discover import service_info_as_text_str
-import socket
-from zeroconf import DNSAddress
+import pytest
+from ouster.cli.plugins.discover import\
+    parse_scope_id, format_hostname_for_url, \
+    get_output_for_sensor, get_text_for_oserror, is_link_local_ipv6_address_and_missing_scope_id
 
 
-SOCKET_TIMEOUT = 2
-FAKESERVER = 'fakeserver.'
+def test_format_hostname_for_url():
+    test_addr_ipv4 = '169.254.169.254'
+    test_addr_ipv6 = '200a:aa8::8a2e:370:1337'
+    test_hostname = "os-122247000785.local"
+    assert format_hostname_for_url(test_addr_ipv4) == test_addr_ipv4
+    assert format_hostname_for_url(test_addr_ipv6) == f'[{test_addr_ipv6}]'
+    assert format_hostname_for_url(test_hostname) == test_hostname
 
 
-class FakeInfo:
-    def __init__(self, fake_server, fake_addresses):
-        self.fake_server = fake_server
-        self.fake_addresses = fake_addresses
-
-    def dns_addresses(self):
-        _type = 0  # not important to us
-        _class = 0  # not important to us
-        _ttl = 0  # not important to us
-        return [
-            DNSAddress(
-                self.fake_server,
-                _type,
-                _class,
-                _ttl,
-                socket.inet_pton(socket.AF_INET6 if ':' in address else socket.AF_INET, address)
-            )
-            for address in self.fake_addresses
-        ]
-
-    def parsed_scoped_addresses(self):
-        return self.fake_addresses
-
-    @property
-    def server(self):
-        return self.fake_server
+def test_parse_scope_id():
+    # It returns a tuple consisting of the ip address string and an optional integer
+    # representing the scope id (if present)
+    test_addr_ipv4 = '169.254.169.254'
+    test_addr_ipv6 = '200a:aa8::8a2e:370:1337'
+    assert parse_scope_id(test_addr_ipv6) == (test_addr_ipv6, None)
+    assert parse_scope_id(f"{test_addr_ipv6}%5") == (test_addr_ipv6, 5)
+    assert parse_scope_id(f"{test_addr_ipv6}%0") == (test_addr_ipv6, 0)
+    assert parse_scope_id(f"{test_addr_ipv6}%0") == (test_addr_ipv6, 0)
+    assert parse_scope_id(test_addr_ipv4) == (test_addr_ipv4, None)
+    # it raises a ValueError if the scope id is not an integer
+    with pytest.raises(ValueError):
+        parse_scope_id(f"{test_addr_ipv6}%invalid")
 
 
-class MockSocket:
-    def __init__(self, af, sock_type, proto):
-        print(f"Constructed MockSocket with {af}, {sock_type}")
-
-    def connect(self, *args):
-        print(f"Called connect with {args}")
-
-    def settimeout(self, timeout):
-        print(f"Called settimeout with {timeout}")
-
-    def close(self):
-        pass
+def test_is_link_local_ipv6_address_and_missing_scope_id():
+    with pytest.raises(ValueError):
+        is_link_local_ipv6_address_and_missing_scope_id("notanaddress")
+    assert not is_link_local_ipv6_address_and_missing_scope_id("10.34.80.17")
+    assert not is_link_local_ipv6_address_and_missing_scope_id("200a:aa8::8a2e:370:1337")
+    assert is_link_local_ipv6_address_and_missing_scope_id("fe80:aa8::8a2e:370:1337")
+    assert not is_link_local_ipv6_address_and_missing_scope_id("fe80:aa8::8a2e:370:1337%2")
 
 
-def mock_socket(addr_family, sock_type, proto):
-    return MockSocket(addr_family, mock_socket, proto)
+def test_text_output():
+    sensor_json = {
+        "active_config": {
+            "udp_dest": "10.34.80.17",
+            "udp_port_imu": 9503,
+            "udp_port_lidar": 9003,
+        },
+        "addresses": [
+            "10.34.26.98"
+        ],
+        "hostname": "os-992343000025.local.",
+        "network": {
+            "hostname": "os-992343000025",
+            "ipv4": {
+                "addr": "10.34.26.98/24",
+                "link_local": "169.254.67.1/16",
+                "override": None
+            },
+            "ipv6": {
+                "link_local": "fe80::be0f:a7ff:fe00:a992/64"
+            },
+        },
+        "sensor_info": {
+            "image_rev": "ousteros-image-prod-aries-v2.5.2+20230714195410",
+            "prod_line": "OS-2-128",
+            "prod_sn": "992343000025",
+        }
+    }
+    text, color = get_output_for_sensor(sensor_json)
+    assert color == 'white'
+    assert "OS-2-128 - 992343000025" in text, text
+    assert "* UDP destination address: 10.34.80.17" in text, text
+    assert "* IPv4 DHCP 10.34.26.98/24" in text
 
 
-def test_service_info_as_text_str(monkeypatch):
-    """It should format correctly even with no addresses in the info."""
-    def mock_get(_, **kwargs):
-        raise RuntimeError("err")
+def test_get_text_for_oserror():
+    e = OSError("Invalid Argument")
+    e.errno = 22
+    with pytest.raises(ValueError):
+        get_text_for_oserror("prefix", "address", e)
 
-    with monkeypatch.context() as m:
-        m.setattr(requests, "get", mock_get)
-        m.setattr(socket, "socket", mock_socket)
-
-        text, color, error = service_info_as_text_str(FakeInfo(FAKESERVER, []), SOCKET_TIMEOUT)
-        server, address, prod_line, dest_ip, lidar_port, imu_port = text.split()
-        assert FAKESERVER == server
-        assert address == '-'
-        assert prod_line == '-'
-
-        fake_addresses = ["192.168.100.200", "200a:aa8::8a2e:370:1337"]
-        text, color, error = service_info_as_text_str(FakeInfo(FAKESERVER, fake_addresses), SOCKET_TIMEOUT)
-        server, address, prod_line, dest_ip, lidar_port, imu_port = text.split()
-        assert FAKESERVER == server
-        assert address == fake_addresses[0]
-        assert prod_line == '-'
-
-
-def test_service_info_as_text_str_2(monkeypatch):
-    """It should format correctly even when sensor metadata can't be retrieved."""
-    def mock_get(_, **kwargs):
-        raise RuntimeError("err")
-
-    with monkeypatch.context() as m:
-        m.setattr(requests, "get", mock_get)
-        m.setattr(socket, "socket", mock_socket)
-
-        fake_addresses = ["192.168.100.200", "200a:aa8::8a2e:370:1337"]
-        text, color, error = service_info_as_text_str(FakeInfo(FAKESERVER, fake_addresses), SOCKET_TIMEOUT)
-        server, address, prod_line, dest_ip, lidar_port, imu_port = text.split()
-        assert FAKESERVER == server
-        assert address == fake_addresses[0]
-        assert prod_line == '-'
-
-
-def test_service_info_as_text_str_3(monkeypatch):
-    """It should set prod_line when the HTTP response contains it."""
-    def mock_get(url, **kwargs):
-        class MockResponse:
-            def json(self):
-                return {'prod_line': 'fake_prod_line'}
-        return MockResponse()
-
-    def mock_config(self, *arg):  # NOQA
-        return None
-
-    with monkeypatch.context() as m:
-        m.setattr("socket.socket", mock_socket)
-        m.setattr(requests, "get", mock_get)
-        m.setattr(ouster.cli.plugins.discover, "_get_config", mock_config)
-
-        fake_addresses = ["192.168.100.200", "200a:aa8::8a2e:370:1337"]
-
-        text, color, error = service_info_as_text_str(FakeInfo(FAKESERVER, fake_addresses), SOCKET_TIMEOUT)
-        server, address, prod_line, dest_ip, lidar_port, imu_port = text.split()
-        assert FAKESERVER == server
-        assert address == fake_addresses[0]
-        assert prod_line == 'fake_prod_line'
-
-
-def test_service_info_as_text_str_4(monkeypatch):
-    """It should not set prod_line when the HTTP response does not contain it."""
-    def mock_get(url, **kwargs):
-        class MockResponse:
-            def json(self):
-                return {}
-        return MockResponse()
-
-    def mock_config(self, *arg):  # NOQA
-        return None
-
-    with monkeypatch.context() as m:
-        m.setattr(requests, "get", mock_get)
-        m.setattr(socket, "socket", mock_socket)
-        m.setattr(ouster.cli.plugins.discover, "_get_config", mock_config)
-
-        fake_addresses = ["192.168.100.200", "200a:aa8::8a2e:370:1337"]
-        text, color, error = service_info_as_text_str(FakeInfo(FAKESERVER, fake_addresses), SOCKET_TIMEOUT)
-        server, address, prod_line, dest_ip, lidar_port, imu_port = text.split()
-        assert FAKESERVER == server
-        assert address == fake_addresses[0]
-        assert prod_line == '-'
-
-
-def test_ipv6_url(monkeypatch):
-    """It should produce the correct url"""
-    test_ipv6_address = "200a:aa8::8a2e:370:1337"
-    test_url = f"http://[{test_ipv6_address}]/api/v1/sensor/metadata/sensor_info"
-
-    def mock_get(url, **kwargs):
-        class MockResponse:
-            def json(self):
-                assert test_url == url
-                return {}
-
-        return MockResponse()
-
-    def mock_config(self, *arg):  # NOQA
-        return None
-
-    with monkeypatch.context() as m:
-        m.setattr(requests, "get", mock_get)
-        m.setattr(socket, "socket", mock_socket)
-        m.setattr(ouster.cli.plugins.discover, "_get_config", mock_config)
-
-        fake_addresses = [test_ipv6_address]
-        _, _, _ = service_info_as_text_str(FakeInfo(FAKESERVER, fake_addresses), SOCKET_TIMEOUT)
+    txt = get_text_for_oserror("prefix", "fe80::be0f:a7ff:fe00:a992", e)
+    from sys import version_info  # TODO: monkeypatch this or make it a fn parameter
+    if version_info < (3, 9):
+        assert "this version of Python does not support scoped link-local IPv6 addresses" in txt
+    else:
+        assert txt == "prefix: Invalid Argument"